+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased (0.5.8)]
+
+### Added
+
+-   `qwertypy.dsa.single_source_shortest_path`
+-   `qwertypy.dsa.all_source_shortest_path`
+
 ## [0.5.7]
 
 ### Updated